+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added `PI_IGNORE_PROJECT_ENV=1` to exclude launch-directory dotenv files while retaining launcher and OMP-level environment sources ([#4599](https://github.com/can1357/oh-my-pi/issues/4599)).
+
 ## [18.1.6] - 2026-09-03
 
 ### Added